+++ src/styled/ThemesStyled.js
@@ -1,0 +1,10 @@
+const theme = {
+    colors: {
+        cSky: '#CAEBF2',
+        cCarbon: '#A9A9A9',
+        cWatermelon: '#FF3B3F',
+        cNeutral: '#EFEFEF',
+        cWhite: '#FFFFFF'
+    }
+}
+export default theme
